Integrated function of Character.cs into PlayerBehavior.cs, also made the health bar change with the actual health, which didn't used to work
