--- Assets/Scripts/Character/HealthAndHeal/Character.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7fe0801354220c745ad26b5a7547924d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
